Found duplicate IT class. Combining and removing duplicate.
